+++ CHANGELOG.md
@@ -13,6 +13,8 @@ after an evil butler tries to steal their inheritance.
 
 - **0.7.8 Currently in development**
   - Errors and Patina Exceptions now appear in red for better readability.
+  - Included the `p-report` command. This will now display a short report on the
+    details of the current system.
 - **0.7.7 Released Tuesday, April 28th 2020**
   - Included new component `patina_system_patina_status.sh` which includes the
     command `p-status`. This command will display a list of Patina-compatible